Add endwise rule for Ruby
I wrote some endwise rules for Ruby. It's not perfect and it's just based on my daily usage, but I think this one can be a good start.
There's one case I cannot manage to do is if things like `if condition` is at the beginning of a line, it won't have endwise, because it conflicts with inline `if condition`. I have to choose one. I appreciate if you guys have any advice on this