fix: change to check ts_node only.
function `is_endwise_node` is not correct.I really want to
check if that node already have a close tag then it will not add an
endwise pairs but seem it is not possible right now.
Because tree-sitter will group a wrong node when it have wrong
syntax(missing end wise pair).