cmp: Fixed double parentheses bug
When commiting a function completion using the '(' character in cmp you
end up with two pairs of parentheses instead of one.

nvim-cmp has just merged a PR that sends the `commit_character` via the
`confirm_done` event.

This commit uses this information to avoid adding a second pair of
parentheses when the user has alredy typed one.