fix: handle primes properly in text filetypes (#345)
In text files (markdown, tex, text, ...) primes should not be considered
as opening single quotes, consider the following example:

    Ben's friends say: '...'

The first single quote should not be considered as an opening single
quote.

In this commit we treat single quotes directly following a word
character as a prime. This only applies to text filetypes.