fix: prevent returned conds lambda from caching the first length parameter
Problem:

I faced this issue when trying to write a simple rule for a block comment:

```lua
      npairs.add_rule(
        Rule('/**', '  */')
        :with_pair(cond.not_after_regex('.-%*/', -1))
        :set_end_pair_length(3)
      )
```

This rule kept failing and I thought I was going insane or just dumb but it turns out due to the odd nature of how variables are captured in lua. The first time that the `not_after_regex` function is run it instantiates the function to be returned with the *cached* value of `length`, and this is essentially always going to be the referenced value from within the returned function. I.e.:

```lua
cond.not_after_regex = function(regex, length)
    length = length or 1
    if not regex then
        return cond.none()
    end
    ---@param opts CondOpts
    return function(opts)
        log.debug('not_after_regex')
        if length < 0 then
            length = #opts.line
            -- LENGTH and #OPTS.LINE ARE NOT ALWAYS EQUAL HERE
        end
        local str = utils.text_sub_char(opts.line, opts.col, length)
        if str:match(regex) then
            return false
        end
    end
end
```

This is what caused my rule to fail. Instantiation a new local variable from within the returned lambda function fixes this issue in my testing, and I implemented it in the other regex related conditions as well. Really hope this can get merged, thank you